Dougdev

+++ backend/tests/core/Entities/TenantHelper.cs
@@ -14,7 +14,7 @@ namespace Pims.Core.Test
         /// <param name="code"></param>
         /// <param name="name"></param>
         /// <returns></returns>
-        public static Entity.Tenant CreateTenant(int id = 1, string code = "TEST", string name = "Test Tenant")
+        public static Entity.Tenant CreateTenant(long id = 1, string code = "TEST", string name = "Test Tenant")
         {
             var tenant = new Entity.Tenant(code, name)
             {
